[bot] Add initial message to messageQueue
This commit adds an initial message to the messageQueue array in the bot.ts file of the MelodyScoutLog_Bot module. The added message is "⚡ - MelodyScoutLog_Bot - Started (First message)".

[botFunctions] Update pool concurrency

This commit updates the concurrency of the PromisePool in the runPlayingnowCallback and runPlayingnowCommand functions in the playingnow.ts and playingnow.ts files, respectively. The concurrency is changed from 20 to 5.

[msGithubApi] Reduce max time after upload

This commit reduces the maximum time after upload in the Files class of the msGithubApi module. The value is changed from 120000 to 70000.

[grammyFunctions] Add error handling

This commit adds error handling to the ctxReply, ctxAnswerInlineQuery, and ctxPinMessage functions in the grammyFunctions.ts file. Errors are caught and logged using the advError function.