fix: Estimate the total playtime for tracks with no duration
Estimate the total playtime for tracks that don't have a duration by using the median track duration of the tracks that do have a duration. This estimate is used when there are more than 150 tracks with no duration. The estimated total playtime is calculated by multiplying the median track duration by the playcount of each track. The estimated playtime is logged for debugging purposes.